:art: Refactored and changed to Typescript
