+++ server/.env.example
@@ -1,0 +1,4 @@
+JWT_SECRET ='m@c-@nd-c#eese'
+MONGODB_URI='mongodb://localhost/springboard'
+PORT=3000
+LOG_LEVEL='debug'
