+++ server/.gitignore
@@ -1,0 +1,8 @@
+.DS_Store
+.env
+!.env.example
+build
+node_modules
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
