+++ server/package.json
@@ -1,0 +1,34 @@
+{
+  "name": "server",
+  "main": "build/index.js",
+  "scripts": {
+    "start": "node build/index.js",
+    "build": "tsc",
+    "lint": "eslint . --ext .ts",
+    "dev": "tsc-watch --onSuccess 'node ./build/index.js'"
+  },
+  "devDependencies": {
+    "@types/express": "^4.17.8",
+    "@types/mongodb": "^3.5.31",
+    "@types/node": "^14.14.5",
+    "@typescript-eslint/eslint-plugin": "^4.6.0",
+    "@typescript-eslint/parser": "^4.6.0",
+    "eslint": "^7.12.1",
+    "tsc-watch": "^4.2.9",
+    "typescript": "^4.0.5"
+  },
+  "dependencies": {
+    "bcrypt": "^5.0.1",
+    "cors": "^2.8.5",
+    "dotenv": "^8.2.0",
+    "express": "^4.17.1",
+    "helmet": "^4.1.1",
+    "jsonwebtoken": "^8.5.1",
+    "mongodb": "^3.6.2",
+    "mongoose": "^6.0.8",
+    "validator": "^13.6.0",
+    "winston": "^3.3.3"
+  },
+  "author": "",
+  "license": ""
+}
