--- server/src/middleware/auth.js
@@ -1,26 +1,0 @@
-const jwt = require("jsonwebtoken");
-const User = require("../models/user");
-
-// creating a authentication function
-const auth = async (req, res, next) => {
-  try {
-    const token = req.header("Authorization").replace("Bearer ", "");
-    const decoded = jwt.verify(token, "thisismynewcourse");
-    const user = await User.findOne({
-      _id: decoded._id,
-      "tokens.token": token,
-    });
-
-    if (!user) {
-      throw new Error();
-    }
-
-    req.token = token;
-    req.user = user;
-    next();
-  } catch (e) {
-    res.status(401).send({ error: "Please Authenticate" });
-  }
-};
-
-module.exports = auth;
