+++ server/src/middleware/auth.ts
@@ -1,0 +1,25 @@
+import jwt from 'jsonwebtoken';
+import { User } from '../models/user';
+import { Request, Response, NextFunction } from 'express';
+
+// creating a authentication function
+export const auth = async (req: Request, res: Response, next: NextFunction): Promise<void> => {
+  try {
+    const token = req.header('Authorization').replace('Bearer ', '');
+    const decoded = jwt.verify(token, 'thisismynewcourse');
+    const user = await User.findOne({
+      _id: decoded._id,
+      'tokens.token': token,
+    });
+
+    if (!user) {
+      throw new Error();
+    }
+
+    req.token = token;
+    req.user = user;
+    next();
+  } catch (e) {
+    res.status(401).send({ error: 'Please Authenticate' });
+  }
+};
